When creating an external traffic network policy, now only use a single existing Otterize network policy to determine the pod selector used by the external traffic network policy as all of the different pod selectors are equivalent.
